requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
async-timeout==4.0.3
    # via redis
cachetools==5.3.2
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
google-api-core[grpc]==2.17.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-redis
google-auth==2.27.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-redis
google-cloud-bigquery==3.17.2
    # via -r requirements.in
google-cloud-core==2.4.1
    # via google-cloud-bigquery
google-cloud-redis==2.15.3
    # via -r requirements.in
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-bigquery
googleapis-common-protos==1.62.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.62.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.1
    # via google-api-core
idna==3.6
    # via requests
packaging==23.2
    # via google-cloud-bigquery
proto-plus==1.23.0
    # via google-cloud-redis
protobuf==4.25.2
    # via
    #   google-api-core
    #   google-cloud-redis
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
python-dateutil==2.8.2
    # via google-cloud-bigquery
redis==5.0.3
    # via -r requirements.in
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
urllib3==2.2.0
    # via requests
